+++ src/app/(main)/builder/appearance/page.tsx
@@ -1,10 +1,6 @@
 'use client';
-import ProfileSection from '../_components/appearance/profile-section';
+import ProfileSection from '../_components/appearance/profile-section/profile-section';
 
 export default function Appearance() {
-  return (
-    <>
-      <ProfileSection />
-    </>
-  );
+  return <ProfileSection />;
 }
