+++ src/app/(main)/builder/links/page.tsx
@@ -1,8 +1,8 @@
 'use client';
 import { DndContext } from '@dnd-kit/core';
 import DragOverlayWrapper from '../_components/drag-overlay-wrapper';
-import AdhocLinks from '../_components/links/adhoc-links';
-import SocialIconsSection from '../_components/links/social-icons-section';
+import SocialIconsSection from '../_components/links/social-links/social-icons-section';
+import AdhocLinks from '../_components/links/adhoc-links/adhoc-links';
 
 export default function Links() {
   return (
